feat(02-01): convert ui.rs to ui/ module with layout.rs, full render(), update main.rs
- Delete airev/src/ui.rs; create airev/src/ui/mod.rs + airev/src/ui/layout.rs
- layout.rs: compute_layout() 3-panel [left,center,right,status_bar] with Spacing::Overlap(1),
  panel_block() with Thick/Plain borders + MergeStrategy::Fuzzy, render_status_bar()
- ui/mod.rs: render(frame, &mut AppState, &Theme) wires layout + panel renders + status bar,
  caches viewport heights before rendering, placeholder List/Paragraph content in panels
- main.rs: add mod app; instantiate AppState::default(), pass &mut state to ui::render()
- Auto-fix [Rule 1 - Bug]: ListState::scroll_down_by/scroll_up_by take u16 not usize in
  ratatui 0.30 (research doc showed usize cast; actual API is u16)

+++ airev/src/app.rs
@@ -133,7 +133,7 @@ impl AppState {
     pub fn scroll_down(&mut self, lines: u16) {
         match self.focus {
             PanelFocus::FileList => {
-                self.file_list_state.scroll_down_by(lines as usize);
+                self.file_list_state.scroll_down_by(lines);
             }
             PanelFocus::Diff => {
                 self.diff_scroll = self.diff_scroll.saturating_add(lines);
@@ -151,7 +151,7 @@ impl AppState {
     pub fn scroll_up(&mut self, lines: u16) {
         match self.focus {
             PanelFocus::FileList => {
-                self.file_list_state.scroll_up_by(lines as usize);
+                self.file_list_state.scroll_up_by(lines);
             }
             PanelFocus::Diff => {
                 self.diff_scroll = self.diff_scroll.saturating_sub(lines);
